[refactor/#154] 마커 추가 및 제거 함수 모듈화
