+++ src/utils/mapFunc.ts
@@ -1,3 +1,4 @@
+import { NavigateFunction } from 'react-router-dom';
 import { MarkerInfoType } from '../types';
 import { IconMarkerMap } from './icon';
 
@@ -88,4 +89,51 @@ export const initMap = (
   mapRef.current = map;
 
   return map;
+};
+
+// 마커 추가 함수
+export const addMarkers = (
+  mapRef: React.RefObject<naver.maps.Map | null>,
+  objectList: naver.maps.Marker[] | null,
+  isClickAble: boolean = false,
+  markerIdList?: number[] | null,
+  navigate?: NavigateFunction
+) => {
+  if (!mapRef.current || !objectList) return;
+
+  objectList.forEach((m: naver.maps.Marker, index: number) => {
+    // 지도에 마커 객체 설정
+    m.setMap(mapRef.current);
+
+    // 이벤트 리스너 (마커 클릭)
+    if (!isClickAble || !markerIdList || !navigate) return;
+    naver.maps.Event.addListener(m, 'click', () => {
+      mapRef.current?.morph(m.getPosition(), 18, {
+        duration: 1000,
+        easing: 'easeOutCubic',
+      });
+
+      const isSame = window.location.pathname.includes(
+        `/map/detail/${markerIdList[index]}`
+      );
+
+      if (!isSame) {
+        navigate(`/map/detail/${markerIdList[index]}`);
+      }
+    });
+  });
+
+  // const clustering = initCluster(markerObjectList, mapRef.current!);
+  /* todo : naver cloud api map forum에서 클러스터별 최상단 마커의 종류에 따른 (클러스터 아이콘) 설정이 가능하다고 답변받을시
+   * clustering.setIcons([clusterIconList[카테고리]])를 사용하여 클러스터 아이콘 지정 구현
+   */
+};
+
+// 마커 제거 함수
+export const deleteMarkers = (objectList: naver.maps.Marker[] | null) => {
+  if (!objectList) return;
+
+  objectList.forEach((m) => {
+    m.setMap(null);
+  });
 };
